Refactor done. Rotate functional and tested.

+++ include/load_map.h
@@ -6,7 +6,7 @@
 /*   By: pecavalc <pecavalc@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/09 15:55:39 by pecavalc          #+#    #+#             */
-/*   Updated: 2025/09/25 10:26:25 by pecavalc         ###   ########.fr       */
+/*   Updated: 2025/09/25 14:47:24 by pecavalc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,9 +18,9 @@ typedef struct s_app	t_app;
 
 typedef struct s_3d_vertex
 {
-	int	x;
-	int	y;
-	int	z;
+	double	x;
+	double	y;
+	double	z;
 }	t_3d_vertex;
 
 typedef struct s_model
